feat(netstring)!: Update packaging for NESM/RESM bridge
BREAKING CHANGE: No longer supports direct use from CommonJS
