+++ packages/netstring/NEWS.md
@@ -1,5 +1,10 @@
 ## Next release
 
+- *BREAKING*: Removes CommonJS and UMD downgrade compatibility.
+  Supporting both Node.js ESM and the `node -r esm` shim requires the main
+  entry point module to be ESM regardless of environment.
+  UMD and CommonJS facets will likely return after all dependees have migrated
+  away from depending upon the `esm` JavaScript module emulator.
 - Fixes a problem with the external visibility of TypeScript types.
 
 ## Release 0.1.0 (2021-04-26)
