+++ packages/netstring/index.d.ts
@@ -1,2 +1,1 @@
-export { netstringReader } from './src/reader.js';
-export { netstringWriter } from './src/writer.js';
+export * from './src/main.js';
