+++ packages/netstring/jsconfig.json
@@ -3,6 +3,7 @@
     "target": "esnext",
     "module": "esnext",
     "noEmit": true,
+    "downlevelIteration": true,
     "strictNullChecks": true,
     "moduleResolution": "node"
   },
