+++ packages/netstring/package.json
@@ -14,47 +14,25 @@
     "url": "https://github.com/endojs/endo/issues"
   },
   "type": "module",
-  "parsers": {
-    "js": "mjs"
-  },
-  "main": "./dist/netstring.cjs",
+  "main": "./index.js",
   "module": "./index.js",
-  "browser": "./dist/netstring.umd.js",
-  "unpkg": "./dist/netstring.umd.js",
   "types": "./index.d.ts",
   "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "import": "./index.js",
-      "require": "./dist/netstring.cjs",
-      "browser": "./dist/netstring.umd.js"
-    },
-    "./reader": {
-      "import": "./reader.js",
-      "require": "./dist/netstring-reader.cjs",
-      "browser": "./dist/netstring-reader.umd.js"
-    },
-    "./writer": {
-      "import": "./writer.js",
-      "require": "./dist/netstring-writer.cjs",
-      "browser": "./dist/netstring-writer.umd.js"
-    }
+    ".": "./index.js",
+    "./reader.js": "./reader.js",
+    "./writer.js": "./writer.js",
+    "./package.json": "./package.json"
   },
   "scripts": {
-    "build": "rollup --config rollup.config.js",
-    "clean": "rm -rf dist",
-    "cover": "nyc ava",
+    "build": "exit 0",
     "lint": "yarn lint:types && yarn lint:js",
-    "lint-fix": "eslint --fix '*.js' '**/*.js' 'test*/**.js'",
-    "lint:js": "eslint '*.js' '**/*.js' 'test*/**.js'",
+    "lint-fix": "eslint --fix '**/*.js'",
+    "lint:js": "eslint '**/*.js'",
     "lint:types": "tsc --build jsconfig.json",
-    "prepublish": "yarn clean && yarn build",
     "test": "ava"
   },
-  "dependencies": {},
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^6.1.0",
+    "@agoric/eslint-config": "^0.3.6",
     "ava": "^3.12.1",
     "babel-eslint": "^10.0.3",
     "eslint": "^7.23.0",
@@ -63,19 +41,15 @@
     "eslint-plugin-eslint-comments": "^3.1.2",
     "eslint-plugin-import": "^2.19.1",
     "eslint-plugin-prettier": "^3.1.2",
-    "nyc": "^15.1.0",
     "prettier": "^1.19.1",
-    "rollup": "1.31.0",
-    "rollup-plugin-terser": "^5.1.3",
     "typescript": "^4.0.5"
   },
   "files": [
     "LICENSE*",
-    "dist",
-    "src",
-    "types",
-    "*.js",
-    "*.d.ts"
+    "index.d.ts",
+    "index.js",
+    "reader.js",
+    "writer.js"
   ],
   "publishConfig": {
     "access": "public"
